Refactor: move `encode_headers` to utils.
